Move constructor of StringArray to base class
